Linux rebuild: added gthread check.
git-svn-id: http://svn.tilp.info/repos/tiemu@168 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/core/core.c
@@ -31,6 +31,7 @@
 #include <sys/types.h>
 #include <stdlib.h>
 #include <string.h>
+#include <unistd.h> // usleep
 
 #include "intl.h"
 #include "struct.h"
@@ -53,7 +54,7 @@
 
 
 G_LOCK_DEFINE(running);
-static volatile byte running = 0;
+static volatile int running = 0;
 
 // run as a separate thread
 gpointer ti68k_engine(gpointer data)
